update wheel definitions; require PyQt6
Currently requiring PyQt6.2 to be installed, since we can't specify
"either 5.x or 6.x" in the requirements. The built wheel should still work
if the user manually installs PyQt5 and force-removes PyQt6.

Whether this becomes permanent or needs reverting will depend on how
many issues are discovered in subsequent testing.